Update README.md
Beverage Survey

Write a program that performs a survey tally on beverages. The program should prompt for the next person until a sentinel value of –1 is entered to terminate the program. Each person participating in the survey should choose their favorite beverage from the following list:

Coffee
Tea
Soda
Orange Juice
Sample Run:

1 - Coffee
2 - Tea
3 - Soda
4 - Orange Juice
-1 - QUIT

Please input the favorite beverage of person #1: Choose 1, 2, 3, or 4 from the above menu or -1 to exit the program.
4
Please input the favorite beverage of person #2: Choose 1, 2, 3, or 4 from the above menu or -1 to exit the program.
3
Please input the favorite beverage of person #3: Choose 1, 2, 3, or 4 from the above menu or -1 to exit the program.
1
Please input the favorite beverage of person #4: Choose 1, 2, 3, or 4 from the above menu or -1 to exit the program.
1
Please input the favorite beverage of person #5: Choose 1, 2, 3, or 4 from the above menu or -1 to exit the program.
2
Please input the favorite beverage of person #6: Choose 1, 2, 3, or 4 from the above menu or -1 to exit the program.
-1

The total number of people surveyed is 5

The results are as follows:

Beverage and Number of Votes
Coffee: 2
Tea: 1
Soda: 1
Orange Juice: 1